chore: refactor code, added functionality for saving state of dashboard
